created and integrated home router
